Modern-C++ memory-safety refactors from full-codebase audit (pass 4)
Full-codebase line-by-line audit (all 360 production source files) for
modern-C++ idioms that eliminate memory-safety risk. As this is the fourth
such pass, the safe surface is small; every change below is behavior-identical
or strictly safer, with no -Wconversion regression, no new branch/function
(coverage-neutral), and no change to serialization/wire-format/RNG/parity.

- gloader.cpp data_copy(): widen the pixie allocation size to std::size_t
  before the product (static_cast<std::size_t>(d.w) * d.h * d.frames) so the
  length feeding make_unique and copy_n is computed in 64-bit arithmetic,
  removing an int-overflow footgun if the per-dimension clamps ever change.
  Matches the already-hardened copy_pixie() in gloader_ctf.cpp.
- view.cpp viewscreen::clear(): replace a hand-rolled 64000-iteration zero
  loop with videobuffer.fill(0), tying the clear to the std::array's real
  extent and dropping the magic size.
- family_mage.cpp: value-initialize possible_specials[] with {} instead of a
  separate zero loop, closing the indeterminate-value window.
- sai2x.cpp: the four destination-pixel stores now use reinterpret_cast<Uint32*>
  instead of C-style pointer casts (matches the rest of the file).
- pixie.cpp / radar.cpp / picker_team_build.cpp: C-style casts -> static_cast.
- video_sdl.cpp: NULL -> nullptr in the test-mode SDL_BlitSurface call.

Verified: ci-test and dev-release builds clean (no warnings); full suite
43/43 with ctest --repeat until-pass:3.

+++ src/resources/gloader.cpp
@@ -401,7 +401,7 @@ PixieData data_copy(const PixieData& d)
     result.w = d.w;
     result.h = d.h;
 
-    size_t len = d.w * d.h * d.frames;
+    size_t len = static_cast<std::size_t>(d.w) * d.h * d.frames;
     result.data = std::make_unique<unsigned char[]>(len);
     std::copy_n(d.data.get(), len, result.data.get());
